requirements.txt
torch>=2.1.0
tqdm>=4.66.1
transformers>=4.30.0
torchvision>=0.15.1
scikit-learn>=1.3.0
Pillow>=10.0.0
wandb>=0.15.0